code/requirements.txt
pyannote.audio
numpy
pyannote.core
pyannotebook
ipywidgets
ipyevents
matplotlib
pandas